Fix conversion when using 64 bit system.
Using 'long' on 64 bit systems will garble the numeric values, hence using
integer here.

+++ conf_scr.c
@@ -51,7 +51,7 @@ void waddstr_item(WINDOW *w, int y, struct conf_item *item, char hilight)
 
 		switch (item->type) {
 			case t_int:
-				sprintf(s, "%ld", *(long int *)item->v);
+				sprintf(s, "%d", *(int *)item->v);
 				break;
 			case t_float:
 				sprintf(s, "%f", *(double *)item->v);
